+++ tests/date/parseDateProperty.spec.ts
@@ -1,92 +1,81 @@
 import {
     CalendarDate,
     CalendarDateTime,
-    parseDateProperty,
     Property,
+    parseDateProperty,
 } from '../../src'
 
 it('returns a CalendarDateTime with value type DATE-TIME', () => {
-    const property: Property = {
-        name: 'DTSTART',
-        params: ['VALUE=DATE-TIME'],
-        value: '20250729T120000Z',
-    }
+    const property = new Property('DTSTART', '20250729T120000Z', {
+        VALUE: 'DATE-TIME',
+    })
     const result = parseDateProperty(property)
     expect(result).toBeInstanceOf(CalendarDateTime)
 })
 
 it('returns a CalendarDate with value type DATE', () => {
-    const property: Property = {
-        name: 'DTSTART',
-        params: ['VALUE=DATE'],
-        value: '20250729',
-    }
+    const property = new Property('DTSTART', '20250729', {
+        VALUE: 'DATE',
+    })
     const result = parseDateProperty(property)
     expect(result).toBeInstanceOf(CalendarDate)
 })
 
 it('returns a CalendarDateTime when no parameters are given', () => {
-    const property: Property = {
-        name: 'DTSTART',
-        params: [],
-        value: '20250729T120000Z',
-    }
+    const property = new Property('DTSTART', '20250729T120000Z')
     const result = parseDateProperty(property)
     expect(result).toBeInstanceOf(CalendarDateTime)
 })
 
-it('returns a CalendarDateTime with no value type is given', () => {
-    const property = {
-        name: 'DTSTART',
-        params: ['OTHER=PARAM'],
-        value: '20250729T120000Z',
-    }
+it('returns a CalendarDateTime when no value type is given', () => {
+    const property = new Property('DTSTART', '20250729T120000Z', {
+        OTHER: 'PARAM',
+    })
     const result = parseDateProperty(property)
     expect(result).toBeInstanceOf(CalendarDateTime)
 })
 
 it('throws when parsing date-time as date', () => {
-    const property: Property = {
-        name: 'DTSTART',
-        params: ['VALUE=DATE'],
-        value: '20250729T120000Z',
-    }
+    const property = new Property('DTSTART', '20250729T120000Z', {
+        VALUE: 'DATE',
+    })
     expect(() => parseDateProperty(property)).toThrow()
 })
 
 it('throws when parsing date as date-time', () => {
-    const property: Property = {
-        name: 'DTSTART',
-        params: ['VALUE=DATE-TIME'],
-        value: '20250729',
-    }
+    const property = new Property('DTSTART', '20250729', {
+        VALUE: 'DATE-TIME',
+    })
+    expect(() => parseDateProperty(property)).toThrow()
+})
+
+it('throws when parsing date as inferred date-time', () => {
+    // DTSTART is DATE-TIME by default
+    const property = new Property('DTSTART', '20250729')
     expect(() => parseDateProperty(property)).toThrow()
 })
 
 it('ignores the property name', () => {
-    const params: string[] = []
     const value: string = '20250729T120000Z'
+    const dateTime = new CalendarDateTime(value)
     const properties: Property[] = [
-        { name: 'DTSTART', params, value },
-        { name: 'DTEND', params, value },
-        { name: 'DTSTAMP', params, value },
-        { name: 'CREATED', params, value },
+        new Property('DTSTART', value),
+        new Property('DTEND', value),
+        new Property('DTSTAMP', value),
+        new Property('CREATED', value),
+        new Property('X-PROPERTY', value),
     ]
 
     const parsed = properties.map(property => parseDateProperty(property))
 
-    parsed.forEach(a => {
-        parsed.forEach(b => {
-            expect(a).toEqual(b)
-        })
+    parsed.forEach(parsedValue => {
+        expect(parsedValue).toStrictEqual(dateTime)
     })
 })
 
 it('throws if value type is illegal', () => {
-    const property: Property = {
-        name: 'DTSTART',
-        params: ['VALUE=BINARY'],
-        value: '20250729T120000Z',
-    }
+    const property = new Property('DTSTART', '20250729T120000Z', {
+        VALUE: 'BINARY',
+    })
     expect(() => parseDateProperty(property)).toThrow()
 })